backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "truepeak-ai"
version = "0.1.0"
description = "AI-powered audio mastering submission and review system"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "pydantic>=2.9.0",
    "pyjwt>=2.9.0",
    "python-multipart>=0.0.12",
    "librosa>=0.10.2",
    "pyloudnorm>=0.1.1",
    "scipy>=1.14.0",
    "numpy>=2.1.0",
    "httpx>=0.27.0",
    "openai>=1.51.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
